+++ common/src/main/java/de/uni_jena/thunibib/ThUniBibCommands.java
@@ -1,7 +1,5 @@
 package de.uni_jena.thunibib;
 
-import de.uni_jena.thunibib.impex.DBTImportIdProvider;
-import de.uni_jena.thunibib.impex.importer.ConfigurableListImportJob;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.Query;
 import org.apache.http.HttpEntity;
@@ -33,7 +31,6 @@ import org.mycore.access.MCRAccessException;
 import org.mycore.backend.jpa.MCREntityManagerProvider;
 import org.mycore.common.MCRConstants;
 import org.mycore.common.MCRException;
-import org.mycore.common.MCRMailer;
 import org.mycore.common.config.MCRConfiguration2;
 import org.mycore.common.content.MCRContent;
 import org.mycore.common.content.MCRJDOMContent;
@@ -49,11 +46,8 @@ import org.mycore.datamodel.metadata.MCRObjectID;
 import org.mycore.frontend.MCRFrontendUtil;
 import org.mycore.frontend.cli.annotation.MCRCommand;
 import org.mycore.frontend.cli.annotation.MCRCommandGroup;
-import org.mycore.services.i18n.MCRTranslation;
 import org.mycore.solr.MCRSolrClientFactory;
 import org.mycore.solr.commands.MCRSolrCommands;
-import org.mycore.ubo.importer.ImportJob;
-import org.mycore.ubo.importer.ListImportJob;
 import org.mycore.ubo.ldap.LDAPAuthenticator;
 import org.mycore.ubo.ldap.LDAPObject;
 import org.mycore.ubo.ldap.LDAPSearcher;
@@ -69,8 +63,6 @@ import javax.naming.ldap.LdapContext;
 import java.io.BufferedInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -83,6 +75,7 @@ import static org.mycore.common.MCRConstants.MODS_NAMESPACE;
 import static org.mycore.common.MCRConstants.XPATH_FACTORY;
 
 @MCRCommandGroup(name = "ThUniBib Commands")
+
 public class ThUniBibCommands {
     private static final Logger LOGGER = LogManager.getLogger(ThUniBibCommands.class);
 
@@ -649,34 +642,5 @@ public class ThUniBibCommands {
         } catch (IOException | JDOMException | MCRAccessException e) {
             LOGGER.error("Could not transform xml", e);
         }
-    }
-
-    @MCRCommand(syntax = "thunibib import {0} from dbt", help = "Imports the object given by its dbt id from DBT")
-    public static void importFromDBT(String dbtid) {
-        LOGGER.info("Importing DBT object {} from dbt", dbtid);
-
-        Element config = new Element("import");
-        config.setAttribute("targetType", "import");
-        config.addContent(new Element("source").setText(dbtid));
-        ListImportJob job = new ConfigurableListImportJob("DBTList", new DBTImportIdProvider());
-
-        try {
-            job.transform(config);
-            job.savePublications();
-        } catch (Exception e) {
-            LOGGER.error("Could not save imported DBT object {}", dbtid, e);
-        }
-        sendMail(job);
-    }
-
-    private static void sendMail(ImportJob job) {
-        String subject = "DBT-Import " + job.getID();
-        String from = MCRConfiguration2.getString("UBO.Mail.From").get();
-        String to = MCRConfiguration2.getString("MCR.Mail.Address").get();
-        String url = MCRFrontendUtil.getBaseURL() + "servlets/solr/select?fq=%2BobjectType%3Amods&q=importID%3A%22"
-            + URLEncoder.encode(job.getID(), StandardCharsets.UTF_8) + "%22";
-        String body = MCRTranslation.translateToLocale("ubo.import.list.email.body", url, "de");
-
-        MCRMailer.send(from, to, subject, body);
     }
 }
