+++ common/src/main/resources/config/thunibib-common/mycore.properties
@@ -34,7 +34,7 @@ MCR.EventHandler.MCRObject.100.Class    = org.mycore.solr.index.MCRSolrIndexEven
 
 # Imports
 MCR.CLI.Classes.External.ThUniBibEnrich = de.uni_jena.thunibib.impex.EnrichmentByAffiliationCommands
-MCR.CLI.Classes.External.ThUniBib       = de.uni_jena.thunibib.his.cli.HISinOneCommands,ThUniBibCommands,de.uni_jena.thunibib.ThUniBibCommands
+MCR.CLI.Classes.External.ThUniBib       = de.uni_jena.thunibib.his.cli.HISinOneCommands,ThUniBibCommands,de.uni_jena.thunibib.ThUniBibCommands,de.uni_jena.thunibib.DBTImportCommands
 
 # PPN2DBT-ID
 MCR.MODS.EnrichmentResolver.DataSource.PPN2DBT-ID.IdentifierTypes = ppn
